+++ src/components/Category/Category.scss
@@ -1,14 +1,19 @@
 button {
-  border: 1px;
-  margin: 10px;
+  font-size: 4px;
+  color: aqua;
 
   .slate {
-    background-color: yellow;
+    background-color: white;
+    border-radius: 5px;
     color: black;
+    font-size: 13px;
   }
-
+  
   .highlighted {
     background-color: black;
+    border-radius: 10px;
     color: white;
+    font-size: 13px;
   }
 }
+
