+++ src/components/Scaffold/Scaffold.css
@@ -4,25 +4,28 @@
   overflow: auto;
   display: flex;
   flex-direction: column;
-  margin-left: 50px;
-  border-width: 1px;
-  border-style: solid;
 }
 
 #scrollableDiv-left {
   background-color: #621f67;
   text-align: left;
-  border-color: #ffffff;
 }
 #scrollableDiv-left h4 {
   color: white;
+  font-size: 35px;
+  padding-left: 20px;
+  padding-top: 30px;
+  font-style: italic;
 }
 
 #scrollableDiv-right {
   background-color: #fed32d;
   text-align: right;
-  border-color: #ffffff;
 }
 #scrollableDiv-right h4 {
   color: #2A3D59;
+  font-size: 35px;
+  padding-right: 20px;
+  padding-top: 30px;
+  font-style: italic;
 }/*# sourceMappingURL=Scaffold.css.map */
