+++ src/utils/constants.js
@@ -1,0 +1,13 @@
+export const CATEGORIES = {
+  'genero': 'Género',
+  'paz': 'Paz',
+  'protesta': 'Protesta',
+  'tributaria': 'Reforma tributaria',
+  'mineroenergetico': 'Energía',
+  'instituciones y democracia': 'Democracia',
+  'seguridad social': 'Seguridad social',
+  'venezuela': 'Venezuela',
+  'corrupcion': 'Corrupción',
+  'educacion': 'Educación',
+  'pobreza y desigualdad': 'Pobreza'
+};
